Move output `.float()` to loss fn and compile it if compiling
ghstack-source-id: 3c96c273a51b2fe2ea03055b670e327a3d0130ea
Pull Request resolved: https://github.com/pytorch/torchtitan/pull/642

+++ torchtitan/models/llama/model.py
@@ -438,7 +438,7 @@ class Transformer(nn.Module):
             h = layer(h, self.freqs_cis)
 
         h = self.norm(h) if self.norm else h
-        output = self.output(h).float() if self.output else h
+        output = self.output(h) if self.output else h
         return output
 
     @classmethod
